output: Move file review model records
The file-review renderer now delegates page selection and layout sizing
while still defining the passive records used by model construction, state
derivation, and rendering.

Those records are shared across several renderer phases, so keeping them
inside output.file_review makes the module open with type definitions
before the reader reaches behavior. That obscures the split between data
shape and output operations.

This commit addresses that by moving the file-review model dataclasses into
output.file_review_model and importing them back into the renderer.

The following tests commit will add a boundary check for the model record
module.
